repo-updater: periodically bump priority of uncloned repositories
The git fetch/clone scheduler is unaware of which repositories are
cloned or uncloned. It is only aware of repositories from the code host
syncer. A repository can not be cloned even if the scheduler is aware of
it for a few reasons:

- The repository failed to cloned.
- The repository was removed due to disk pressure.

This commit adds a periodic sync between the list of repositories
gitserver has cloned and the state of the scheduler. We walk the list of
all repositories in the scheduler and if it is not cloned, we update its
"due" date to be as if the repository was new (default 45s until we
clone it).

Note: all repositories returned from the repo store are in the scheduler
via the code host syncer.