repo-updater: remove updateScheduler mutex (#11632)
As far as I can tell this mutex serves no purpose. The only possible
purpose is to serialize calls to UpdateFromDiff. However, the individual
upsert and remove calls on the scheduler and queue have there own
mutexes.

Possibly if we had two concurrent `UpdateFromDiff` we could have an
inconsistent state, but it would converge at the next UpdateFromDiff
call. Additionally we only have one call site for UpdateFromDiff, and
that never runs concurrently.