zoekt: treat repo names case sensitively (#11670)
Previously we treated zoekt repositories as if the case would not match
what is returned from the repo database. However, Zoekt requires the
correct caseing of the repo name to do the search. Additionally Zoekt
uses the correct casing of the repo name when indexing.

So the uses of ToLower when checking if a repo was indexed was
unnecessary. Additionally in the case a repository name changed case we
would incorrectly believe we have indexed the repo, but the actual zoekt
search would fail due to the index being case sensitive.

Additionally avoiding the calls to ToLower should improve
performance. However, I didn't validate this (but naively it should).