repo-updater: populate git fetch scheduler on startup
The git fetch scheduler discovers what repos to fetch/clone by reading
the Synced and SubsetSynced channels from the Syncer. On repo-updater
startup it will only start running once the first Sync is complete (this
is when it first discovers repos). On large installations a sync can
take a while. This change allows the scheduler to start syncing the
pre-existing repos (common case) on startup.