repo-updater: remove streaming syncer feature flag (#11690)
It has defaulted to on for many releases.

Additionally I noticed a bug. If SyncedSubset is nil we should not do
streaming inserts. The reason is Synced will never have the Diff.Added
field populated since it will always be via subset synced. So this would
lead to the repo not being noticed in the git update scheduler. However,
we always have SubsetSynced set (except in tests). So this was a benign
bug.