Add types and store methods for CampaignSpec and ChangesetSpec
The schema is:

- a `campaign` can have one `campaign_spec`
- a `campaign_spec` can have many `changeset_spec`

Most notable is that `changeset_specs.campaign_spec_id` is nullable so
that users can create `ChangesetSpecs` individually and then attach them
to `CampaignSpecs`.

And both tables have a new `rand_id` column that contains a short,
unguessable text ID that we can use as an external GraphQL node ID for
both.