search: factor out symbol search code (#11712)
I was trying to change this function and it was quite hard to
understand. This was mostly due to the symbol search code that was
added. This commit extracts the logic of how we create symbol resolvers
from a zoekt.FileMatch. This also makes it clearer that baseURI and the
commit resolver are only used by symbol search. We now avoid allocating
those values in the more common text search.