zoekt: handle multiple branches in zoektIndexedRepos (#11709)
This teaches zoektIndexedRepos to handle branches other than HEAD. It
uses the same logic in zoektSingleIndexedRepo to assign a Sourcegraph
revspec to a zoekt branch. If we have indexed all the revspecs we want
to search in repo we pass the query on to zoekt.

Since we use the same logic, we also remove zoektSingleIndexedRepo
and just use the implementation in zoektIndexedRepos.

Note: The common case is to search only HEAD in all repos. This code
change should be as performant, except we will now allocate a branches
slice per repo. If this shows up in profiles we can optimize this. (eg a
global var justHEAD = []string{"HEAD"} to reuse).

Note: We need to update the function to return the resolves branches to
search in Zoekt. This will be done in a later commit.

Note: This function will allow us to remove zoektSingleIndexedRepo. This
will be done in a later commit.