graphqlbackend: treat yes/true and no/false values the same
This commit updates the yesNoOnly parser to not return different values
for true and false. Instead it will return yes or no respectiveley. This
simplifies read sites since we don't need to check for true or yes, etc.

This also highlighted a weird bit of logic. Previously if you did
"fork:yes" we would include forks. However, if you did "fork:true" we
only searched forks. I am unsure if this is intentional behaviour, but
it is surprising to me. Rather if a user only wants to search forks I'd
expect them to search "fork:only".