repo-updater: avoid concurrent access to fakes
Now that we do the update in the background, we have concurrent writes
and reads between the test and the syncer. Instead of making our clock
and store fakes safe for concurrent use, we avoid the concurrent
reads.

In the case of already exist we don't need to assert on the state of the
store, just on the repo lookup result. This is a concurrent write due to
how transactions are implemented in our testing fake (even though the
update is a noop).

In the case of clock we use a new version of it per test case.