repo-updater: Do not block in repoLookup for known repos (#11780)
If we have the entry in the DB already, we do not need to block the
repoLookup request. For example GitHub has degraded services today, and
requests for loading repositories timed out on sourcegraph.com due to
this code path. However, if we already have repo metadata we can just
return that in repoLookup. We still do what we did before, but in a
background goroutine. We still need to do it so we can detect metadata
changes (eg renames).