doc/admin: clarify that upgrades require 10m of downtime by default (#11796)
Prior to this commit we advertised that HA cluster deployment upgrades
were zero-downtime. In practice, this is not true as database migrations
may mean that e.g. some frontend replicas during a rolling updates will
be unable to write (and sometimes read) from the database. In practice
this is most likely a service-degradation situation, not ocmplete downtime,
but it is good to be explicit about planning a 10m downtime window.

This change was discussed and approved live with @efritz @tsenart @pecigonzalo @beyang