git: Command retrying made a copy that was never used
commandRetryer is supposed to perform its first attempt without trying
to resolve missing revisions. The previous code made a copy with removed
fields to do this, but didn't actually use the copy.

In practice this meant that when a command was run against a missing
revisions with a supplied value in the EnsureRevisions field we actually
tried to fetch it twice from the remote.