gitserver: always set GIT_DIR
This adds a helper to our GitDir type to update a command such that the
git command will use GitDir. This also helps us avoid casting GitDir to
a string all over the place.

I tried to think up a nicer API, but it started to get tricky. This is
the best I could come up with for now.