search: unify calls to zoektIndexedRepos and zoektSearchHEAD (#11787)
Currently the clients of zoekt need to first call zoektIndexedRepos,
make some adjustments based on arguments, then call zoektSearchHEAD. We
abstract out these details into a single structure indexedSearchRequest.

This also makes it possible for us to maintain more state between the
call from zoektIndexedRepos to zoektSearchHEAD. The main need for this
is to maintain which branch is for which repository revision when
interpreting results from Zoekt. This will be done in another PR.

Note all the duplication between symbols and text search. This was
mostly removed in this commit.