search: directly consult indexed map for default repos (#11836)
This avoids a bunch of needless allocations and simplifies the
implementation of default repos on sourcegraph.com.

Additionally it avoids the use of zoektIndexedRepos which needs to
change to support multiple branches.

Note: we no longer log unindexed repos. This would be easy to
support. However, no one ever read these logs before.