cloud: Repo-updater store repositories clone status in repo table (#11867)
Modifies repo-updater so that it stores the list of cloned repositories in the repo table. It adds a new cloned column to the repo table and a method named SetReposCloned to the repo-updater store, which periodically updates that column for every repo. This allows caching the list of cloned repositories for easy access.