search: register repogroup routes in frontend
This is to prevent us 404ing when loading a repogroup page. This is
copying the hack in the webapp currently and is not sufficient for when
we roll this out. Instead the repogroups should be living under a
subpath, rather than top-level paths.

Additionally the current logic in the webapp is just a string contains,
so registers many more paths. From testing thankfully our frontend
prefers paths that match repos.