search: only search requested indexed branches (#11952)
Previously our indexed search code was updated to detect which branches
are indexed, and if indexed search those repositories. However, we
hadn't yet added code which told Zoekt which branches to search.

This commit records the mapping between the Zoekt branch and the
revision specified by the user. This allows us to tell Zoekt which
branches to search as well as map back Zoekt results to the user
specified revision.

Historical Context: Technically it would be more correct to link to the
SHA and easier to implement. However, that would "lose" your current
browsing context.