search: return branches to index to Zoekt
If we have version contexts enabled we will tell Zoekt to index the
branches in the version context. There is a related Zoekt change to pass
in the repository as well as understand the Branches field being set.

Previously Zoekt was responsible for resolving HEAD. Now that we will do
more than HEAD, it was easier to implement the resolving logic in
Sourcegraph. This will both simplify the responsiblilities of
zoekt-sourcegraph-indexserver as well as allow us to optimize patterns
around resolving version context revisions.