searcher: compile pattern before search (#12144)
It seems CI can have a lot of contention leading to timeouts on the
store even though the store should be instant for
TestSearch_badrequest. We could bump "FetchTimeout", and this could be a
strategy to use if we notice more flakiness in searcher tests. For now
we avoid extra work of speaking to the store for invalid patterns.