Turn ChangsetEvents.Labels into ComputeLabels function
I think it makes more sense to follow the same style for all of these
"compute a state based on the changeset and its events"
methods/functions. `(*ChangesetEvents).Labels` was the last one that we
didn't convert.

Since I've been looking at it in #12198 I decided to change it to a
`ComputeLabels` function (that also does the sorting)