dev: ignore changes to _test.go files (#12251)
We don't need to recompile if a _test.go file is changed since they
don't contribute to main. This change updates watchman to exclude
triggering updates in that case. We only do this for watchman since the
other file watchers do not make it as easy to express these sorts of
rules. I also want to one day make watchman the only way to do this.