search: return a result per branch in indexed results (#12249)
If a file is the same in two branches and we have results on it, our
indexed search currently will only return one result. This would be
misleading in the UI since the user would not know the other branches
have results.

This change will copy how we do it for non-indexed search and just
duplicate results (but with different input revs). In future we should
dedup results and display this information more compactly.