search: always search indexed branches if available (#12154)
This allows our search layer to search both indexed and unindexed
revisions for a repository at the same time.

A user needing to do this is a bit of a rare case outside of
version contexts. And now that we index version context branches
this will only happen when explicitly opting into multiple
branches. There can be some weird reporting artefacts (eg a repo
being listed as both indexed and unindexed search). But better to
have indexed search than not.