codeowners: remove cloud from shared files (#12244)
Cloud currently is the codeowner of many files that are modified by
people all over Sourcegraph. As such our whole team is added to code
reviews where our input really isn't required.

There is discussions about removing code owners entirely. This is a step
in that direction for the more egregerious examples of it for cloud. If
you thought it was important the reviewed you did on something I removed
ownership on, please add your own name (instead of team) to the
corresponding directory.