src-expose: do not run git update-server-info in request path
We have list-repos request timing out. From reading the git
update-server-info sourcecode this operation can be slow:

- It does a git for-each-ref and loads the object.
- It loads all packs.

This feels quick, but I am assuming it can slow down quite a lot for a
large collection of repositories. So my assumption is this is what is
slow in src-expose. This commit moves it into the background (except for
new repos) and ensures only one is running at a time. This shouldn't
need to be run on each request, we do it just in case the hook didn't
run.

If list-repos endpoint is still slow then turning on debug output with
src-expose will help understand which files it is looking at. This is a
educated guess at a fix.