serve-git: port src-expose serve to hidden sub command (#250)
This initial commit copy pastes the serve code from the Sourcegraph repo
at commit d0d94c90. It's mostly copy paste, but we remove some
side-effects we no longer needed to rely on.

I'd like to continue developing this against master and when it is more
ready I'll document it in the CLI and other docs.