search: limit indexed branches per repository to 64
Zoekt only supports indexing up to 64 repositories. This commit enforces
this in the JSON schema for "search.index.branches". Enforcing this for
Version Contexts will require adding a custom validator, which I'll
leave for a later improvement.

Since we don't enforce this everywhere, the options we generate will
deterministically only return up to 64 branches. If an admin specifies
more, it is likely better to index 64 branches than index none and fail.