zoekt: indexserver removes repos from queue
This includes several changes to zoekt indexserver which improve
observability as well as make us behave better during rebalancing. In
particular we would try to index all repos on a replica if we ever
rebalanced. We will still do that, but as soon as the replica comes back
up we will stop doing that.

- 3e5f0f3 indexserver: track last index state as a metric
- cfdfeff indexserver: metrics for queue
- a453878 indexserver: stop indexing repositories if not listed