search: trace and observe each zoekt host
This commit moves tracing and observability from being around the
aggregated zoekt client to being both the aggregated client and a trace
per zoekt replica.

To do this we adjust the prometheus metric to have a hostname label. We
also change the category / family name to indicate if its a search
against a specific zoekt replica or its the aggregation.