search: granular logging of zoekt opts and stats
Previously we used log.Object but this makes it hard to read the logs
since it just does fmt.Sprintf("%v") on the object. We now explicitly
list out each field. This should have the benefit of being more
efficient since we avoid reflection and store the values as ints instead
of strings.