search: specify branch for zoektSearchHEADOnlyFiles
Structural search expects to only search HEAD (currently). However, it
was using a RepoSet which will return results for any branch we have
indexed. I believe this would result in potential benign false positives
for files to search. We switch to instead passing in the computed
RepoBranches (which should be limited to HEAD).