search: handle ctx without deadline
In theory if zoekt is given a context without a deadline (but with
UseFullDeadline true) it will create a goroutine which blocks
forever. To prevent that we can stop the goroutine when the new context
is done.

In practice this shouldn't happen since ctx will have a deadline if
UseFullDeadline is true.