Remove wrapper UpdateChangesets/CreateChangesets (plural) methods
Before we loosened the constraints on the `changesets` table these
methods did batch-insert and batch-update in a single database query.
Since that's not needed (nor that easy) anymore, we can remove these
methods, since they only provide syntactic sugar for a few usecases.

If we decide we need the syntactic sugar back, we can re-add them.