search: clicking a search suggestion no longer appends filters (#12721)
Search suggestions already come with all the existing filters added, so there is no need to add them to the URL when generating the link.

Also fixes a small bug that caused filters to not be toggled properly (eg. after clicking on "lang:c++" and then on "lang:c", the result would be "++" instead of "lang:c++ lang:c". 

Also added unit tests and an integration test for these bugs.