monitoring(frontend): use ratios instead of hard thresholds (#12756)
In general, some of the noisiest alerts in #alerts-cloud are those alerts with hard thresholds, ie "Y+ errors in X minutes" - on larger instances like Sourcegraph Cloud, this could mean we fire alerts on issues that only affect a very small number of users.

This change tries to convert all of the frontend service's hard threshold alerts to ratio-based alerts. We opt to just change them all for consistency. To try and account for smaller instances where a few errors in a few requests can trigger an alert, we add a "for" parameter to each alert, so that requests should consistently error out before causing an alert to be raised.