repo-updater: No private repos on sourcegraph.com (#12908)
This commit makes it so that repo-updater's repo lookup endpoint returns
a "not found" for any private repo in either github.com or gitlab.com.

This is better than only relying on the access token in the external
service not having access to private repos and failing to clone those
repos in gitserver.

This change does not affect the normal syncer code paths, so we'll have
to consider how to not support private code (for now) there, before we
roll out everything from #12699.

Additional context: https://sourcegraph.slack.com/archives/C07KZF47K/p1597091562260500