store: log when failing to fetch archive (#12970)
Currently if we fail we return the error in the response to the
user. However, in most cases a store open fails once the user request
has gone away, so the error is never shown anywhere.

This should help debug errors related to the archive not being available
to search.