web: Add codehost entry for src serve-git (#12754)
"src serve-git" is the new and streamlined version of src-expose. We add
an entry for it in our frontend for code host setup to simplify use of
it.

Additionally this PR updates documentation around src-expose to instead
point to or be about "src serve-git". Note: we still haven't done a
tagged release of src including serve-git. I will ensure that is done
before merging this document so the instructions are valid.

Alternatively we considered add a new code host type. But that is quite
a heavyweight approach from a backend perspective.