campaigns: update a migration with an UPDATE (#13030)
The boat burning migration that was merged in #13001 didn't account for the
possibility that a campaign may have been created in the narrow time window
between the campaign_spec_id and last_updated_at fields being added. Since the
sum total of people who might be affected is likely three, and they all work
here, we'll deal with this by updating the existing migration.