Use a ctx with actor when closing open changesets
This is something I discovered yesterday: when we close the open
changesets after closing a campaign or applying a new spec, we don't
pass in a context that contains the actor.

But the actor is used to check for repository permissions in
`CloseOpenChangesets`.

Result was: all changesets would've been closed, even if you didn't have
repository permissions.